+++ docker/deploy.hyper.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+# Pull the latest
+hyper pull karllhughes/feedbin-mailer