+++ docker/runner.bash
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-set -o allexport; source /app/.env; set +o allexport; /usr/local/bin/node /app/run.js